backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.10.0
pydantic-core==2.27.0
email-validator==2.2.0
PyJWT==2.9.0
argon2-cffi==23.1.0
cryptography==42.0.5
pytest==7.4.3
pytest-mock==3.12.0
httpx==0.25.2
